requirements.txt
absl_py
transformers
torchao
torchdata>=0.8.0
datasets>=2.21.0
tomli=1.1.0
tensorboard
tiktoken
blobfile
tabulate
wandb
